fix(inbound): harden intercom sync + recurrence dedup from full-branch code review
Fixes surfaced by a 5-agent adversarial review of the branch, confidence-
scored and filtered:

Data integrity (poll.go):
- Transient detail/ingest failures now stop the tick BEFORE the failed
  conversation (outcomeRetry) instead of advancing the watermark past it
  or permanently ingesting a summary-only snapshot under its idempotency
  key; permanent per-conversation conditions still degrade to summary.
- Early stops (budget/rate-floor/transient/ctx) step the watermark back
  one second so unprocessed conversations sharing the boundary second
  are re-covered next tick (re-fetches dedup by idempotency key).
- queryEnd drops the +1 so updated_at == now lands next tick, matching
  the documented exclusive upper bound.
- SyncStats.ConversationsSynced accumulates on every exit path — budget
  exhaustion is the common backfill path and silently dropped counts.
- pollError is now set on transient degradation so operators see
  last_error instead of a healthy-looking source.
- Removed the fictitious disabled return contract on fetchAndIngest.

Recurrence signal (feedback_search.go / feedback_similar.go / console):
- FindSimilarFeedback collapses snapshots of the same conversation/
  ticket (evolution capture rows) to their best hit so a long-running
  thread never counts as "recurred N times" against itself.
- GET /feedback/{id}/similar now returns anchor_linked_requests; the
  candidate card prefers the anchor's own links, hides the redundant
  link action when the anchor is already linked, and never offers a
  duplicate promote for an already-tracked anchor.
- firstLine truncates by runes (CJK-safe).

Attribution (attribution.go):
- FeedbackSourceMetaTx now runs inside the per-row savepoint and logs
  failures — a failed SELECT can no longer poison the promote tx.

Hardening / correctness:
- intercomclient.SetTestBaseURL → SetAPIBaseURL; intercom.api_base_url
  is refused under profile: production (it bypasses the *.intercom.io
  allowlist; the egress dial policy still applies in dev).
- Company.MonthlySpend is float64 — fractional monthly_spend no longer
  fails the whole GetCompany decode and drops revenue context.
- enrichWithSyncStats: last_synced_ticket_id only when > 0 (Zendesk-
  only concept); backfill_done surfaces even for empty-window backfills.
- CHANGELOG: moved the misplaced ### Fixed heading that had reclassified
  the Zendesk Added entries; corrected stale doc comments (0.78
  threshold rationale, zendesk meta-key vocabulary, backoff schedule).

+++ internal/inbound/adapter/intercom/fixture_test.go
@@ -193,7 +193,7 @@ func verifyFixtureMeta501(t *testing.T, meta map[string]any) {
 		t.Errorf("custom_attributes = %v", meta["intercom_custom_attributes"])
 	}
 	// Company profile resolved via GET /companies/{id} (revenue context).
-	if meta["intercom_company_monthly_spend"] != 1200 {
+	if meta["intercom_company_monthly_spend"] != float64(1200) {
 		t.Errorf("company_monthly_spend = %v", meta["intercom_company_monthly_spend"])
 	}
 	if meta["intercom_company_plan"] != "Pro" {
